+++ src/models/GetPracticeTeamResponseMembers.ts
@@ -13,13 +13,13 @@
  */
 
 import { mapValues } from "../runtime";
-import type { ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue } from "./ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue";
+import type { ListCatalogItemsResponseDataInnerMedicationGroupOfferCount } from "./ListCatalogItemsResponseDataInnerMedicationGroupOfferCount";
 import {
-  ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueFromJSON,
-  ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueFromJSONTyped,
-  ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueToJSON,
-  ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueToJSONTyped,
-} from "./ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue";
+  ListCatalogItemsResponseDataInnerMedicationGroupOfferCountFromJSON,
+  ListCatalogItemsResponseDataInnerMedicationGroupOfferCountFromJSONTyped,
+  ListCatalogItemsResponseDataInnerMedicationGroupOfferCountToJSON,
+  ListCatalogItemsResponseDataInnerMedicationGroupOfferCountToJSONTyped,
+} from "./ListCatalogItemsResponseDataInnerMedicationGroupOfferCount";
 
 /**
  *
@@ -29,22 +29,22 @@ import {
 export interface GetPracticeTeamResponseMembers {
   /**
    *
-   * @type {ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue}
+   * @type {ListCatalogItemsResponseDataInnerMedicationGroupOfferCount}
    * @memberof GetPracticeTeamResponseMembers
    */
-  total: ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue;
+  total: ListCatalogItemsResponseDataInnerMedicationGroupOfferCount;
   /**
    *
-   * @type {ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue}
+   * @type {ListCatalogItemsResponseDataInnerMedicationGroupOfferCount}
    * @memberof GetPracticeTeamResponseMembers
    */
-  active: ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue;
+  active: ListCatalogItemsResponseDataInnerMedicationGroupOfferCount;
   /**
    *
-   * @type {ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue}
+   * @type {ListCatalogItemsResponseDataInnerMedicationGroupOfferCount}
    * @memberof GetPracticeTeamResponseMembers
    */
-  disabled: ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValue;
+  disabled: ListCatalogItemsResponseDataInnerMedicationGroupOfferCount;
 }
 
 /**
@@ -71,18 +71,9 @@ export function GetPracticeTeamResponseMembersFromJSONTyped(
     return json;
   }
   return {
-    total:
-      ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueFromJSON(
-        json["total"],
-      ),
-    active:
-      ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueFromJSON(
-        json["active"],
-      ),
-    disabled:
-      ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueFromJSON(
-        json["disabled"],
-      ),
+    total: ListCatalogItemsResponseDataInnerMedicationGroupOfferCountFromJSON(json["total"]),
+    active: ListCatalogItemsResponseDataInnerMedicationGroupOfferCountFromJSON(json["active"]),
+    disabled: ListCatalogItemsResponseDataInnerMedicationGroupOfferCountFromJSON(json["disabled"]),
   };
 }
 
@@ -99,17 +90,8 @@ export function GetPracticeTeamResponseMembersToJSONTyped(
   }
 
   return {
-    total:
-      ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueToJSON(
-        value["total"],
-      ),
-    active:
-      ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueToJSON(
-        value["active"],
-      ),
-    disabled:
-      ListCatalogItemsResponseDataInnerPrescriptionRequirementsAllowedQuantitiesInnerValueToJSON(
-        value["disabled"],
-      ),
+    total: ListCatalogItemsResponseDataInnerMedicationGroupOfferCountToJSON(value["total"]),
+    active: ListCatalogItemsResponseDataInnerMedicationGroupOfferCountToJSON(value["active"]),
+    disabled: ListCatalogItemsResponseDataInnerMedicationGroupOfferCountToJSON(value["disabled"]),
   };
 }